requirements.txt
flask
pymysql
sqlalchemy
python-dotenv
b2sdk
mysql-connector-python
